Implemented initialization of dash::util::Config from environment variables

+++ dash/src/util/BenchmarkParams.cc
@@ -117,7 +117,7 @@ void BenchmarkParams::print_header()
   print_param("NUMA domains/node", numa_nodes);
   print_param("CPU max MHz",       cpu_max_mhz);
   print_param("CPU min MHz",       cpu_min_mhz);
-  for (int level = 0; level < cache_sizes.size(); ++level) {
+  for (size_t level = 0; level < cache_sizes.size(); ++level) {
     auto cache_kb     = cache_sizes[level] / 1024;
     auto cache_line_b = cache_line_sizes[level];
     std::ostringstream cn;
